# API-SPECIFICATION.yml
openapi: 3.0.3
info:
  title: Payment Gateway API - Enterprise Payment Processing
  description: |
    # 🚀 Payment Gateway API v1.0
    
    **Enterprise-grade payment processing API** with comprehensive Authorize.Net integration, designed for secure, scalable, and compliant financial transactions.
    
    ## 🎯 Core Capabilities
    
    ### 💳 Payment Processing
    - **Direct Purchase**: One-step authorization and capture
    - **Two-Step Processing**: Separate authorization and capture
    - **Transaction Management**: Void and refund operations
    - **Payment Method Validation**: Pre-transaction verification
    
    ### 📄 Subscription Management
    - **Recurring Billing**: Automated subscription processing
    - **Plan Management**: Flexible subscription plans with trials
    - **Lifecycle Control**: Pause, resume, and cancel subscriptions
    - **Dunning Management**: Intelligent failed payment handling
    
    ### 📊 Analytics & Reporting
    - **Real-time Dashboards**: Live transaction monitoring
    - **Business Intelligence**: Revenue and performance analytics
    - **Compliance Reports**: PCI DSS, GDPR, and SOX reporting
    - **Data Export**: Comprehensive data export capabilities
    
    ### 🔔 Webhook Integration
    - **Real-time Notifications**: Instant transaction updates
    - **Event Processing**: Comprehensive event handling
    - **Retry Logic**: Guaranteed delivery with exponential backoff
    
    ## 🔐 Security & Authentication
    
...
    
...
    
...
    
...
servers:
...
security:
  - BearerAuth: []
  - ApiKeyAuth: []
paths:
...
  /auth/login:
    post:
      tags:
        - 🔐 Authentication
      summary: User Login - Obtain JWT Token
      description: |
        **Authenticate user credentials and receive JWT tokens**
        
        This endpoint validates user credentials (email/username + password) and returns:
        - **Access Token (JWT)**: Short-lived token for API access (15 minutes)
        - **Refresh Token**: Long-lived token for obtaining new access tokens (30 days)
        - **User Profile**: Basic user information and permissions
        
        ### Features
        - 🔒 Secure password validation with bcrypt
        - 🛡️ Brute force protection (5 failed attempts = 30min lockout)
        - 📱 Multi-factor authentication support
        - 🌍 Geographic access control
        - 📊 Login audit trail with IP tracking
        
        ### Rate Limiting
        - **Standard Users**: 5 attempts per 5 minutes
        - **After Failure**: Progressive backoff (1min → 5min → 15min → 30min)
        
      operationId: authenticateUser
      security: []
      requestBody:
        required: true
        description: User credentials for authentication
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            examples:
              email_login:
                summary: Login with Email
                value:
                  email: john.doe@example.com
                  password: SecurePass123!
                  rememberMe: true
                  deviceFingerprint: "device_abc123"
              username_login:
                summary: Login with Username
                value:
                  username: johndoe
                  password: SecurePass123!
                  mfaToken: "123456"
                  ipAddress: "192.168.1.1"
      responses:
        '200':
          description: ✅ Authentication successful
          headers:
            X-Rate-Limit-Remaining:
              description: Number of requests remaining in current window
              schema:
                type: integer
            X-Rate-Limit-Reset:
              description: Unix timestamp when rate limit resets
              schema:
                type: integer
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
              examples:
                successful_login:
                  summary: Successful Login Response
                  value:
                    success: true
                    accessToken: "eyJhbGciOiJSUzI1NiIsInR5cCI6IkpXVCJ9..."
                    refreshToken: "rt_abc123def456..."
                    expiresIn: 900
                    tokenType: "Bearer"
                    user:
                      id: "usr_123456"
                      email: "john.doe@example.com"
                      firstName: "John"
                      lastName: "Doe"
                      roles: ["USER", "PAYMENT_PROCESSOR"]
                      permissions: ["payments:read", "payments:create"]
                      lastLoginAt: "2025-01-04T14:30:00Z"
                      mfaEnabled: true
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          description: 🚫 Authentication failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                invalid_credentials:
                  summary: Invalid Credentials
                  value:
                    success: false
                    error:
                      code: "INVALID_CREDENTIALS"
                      message: "Invalid email or password"
                      details: "Please check your credentials and try again"
                      correlationId: "corr_123456"
                      retryAfter: 60
                account_locked:
                  summary: Account Locked
                  value:
                    success: false
                    error:
                      code: "ACCOUNT_LOCKED"
                      message: "Account temporarily locked due to multiple failed attempts"
                      details: "Please try again in 30 minutes or contact support"
                      lockExpiresAt: "2025-01-04T15:00:00Z"
        '429':
          $ref: '#/components/responses/RateLimited'
...
  /payments/purchase:
    post:
      tags:
        - Payment Processing
      summary: Process a purchase transaction
      description: |
        Process a direct purchase transaction (authorization + capture in one step).
        This immediately charges the customer's payment method and is suitable for most e-commerce transactions.
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/CorrelationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PurchaseRequest'
            example:
              amount: 99.99
              currency: USD
              paymentMethod:
                type: CREDIT_CARD
                cardNumber: "4111111111111111"
                expiryMonth: 12
                expiryYear: 2025
                cvv: "123"
                cardholderName: John Doe
              customer:
                email: john@example.com
                firstName: John
                lastName: Doe
                phone: "+1-555-123-4567"
              billingAddress:
                line1: 123 Main St
                city: New York
                state: NY
                postalCode: "10001"
                country: US
              order:
                orderNumber: ORD-001
                description: Test purchase
      responses:
        '200':
          description: Payment processed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/PaymentError'
        '429':
          $ref: '#/components/responses/RateLimited'
  /payments/authorize:
    post:
      tags:
        - Payment Processing
      summary: Authorize a payment
      description: |
        Process an authorization-only transaction that reserves funds on the customer's
        payment method without capturing them. Use capture endpoint to complete the transaction.
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/CorrelationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthorizeRequest'
      responses:
        '200':
          description: Payment authorized successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/PaymentError'
  /payments/capture:
    post:
      tags:
        - Payment Processing
      summary: Capture an authorized payment
      description: |
        Capture funds from a previously authorized transaction. The authorization
        must be valid and not expired (typically within 30 days).
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/CorrelationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CaptureRequest'
      responses:
        '200':
          description: Payment captured successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/PaymentError'
  /payments/void:
    post:
      tags:
        - Payment Processing
      summary: Void an authorized payment
      description: |
        Cancel a previously authorized transaction before it's captured.
        This releases the hold on the customer's payment method.
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/CorrelationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VoidRequest'
      responses:
        '200':
          description: Payment voided successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/PaymentError'
  /payments/refund:
    post:
      tags:
        - Payment Processing
      summary: Process a refund
      description: |
        Process a full or partial refund for a previously captured transaction.
        Refunds can be processed multiple times up to the original transaction amount.
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/CorrelationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefundRequest'
      responses:
        '200':
          description: Refund processed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/PaymentError'
  /payments/{transactionId}:
    get:
      tags:
        - Payment Processing
      summary: Get transaction status
      description: Retrieve the current status and details of a payment transaction.
      parameters:
        - name: transactionId
          in: path
          required: true
          description: Unique transaction identifier
          schema:
            type: string
            example: txn_1234567890
      responses:
        '200':
          description: Transaction details retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionDetails'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /subscriptions:
    post:
      tags:
        - Subscription Management
      summary: Create a new subscription
      description: |
        Create a new subscription for recurring billing. Supports trial periods,
        custom billing cycles, and various subscription plans.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSubscriptionRequest'
      responses:
        '201':
          description: Subscription created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubscriptionResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'

    get:
      tags:
        - Subscription Management
      summary: List subscriptions
      description: Retrieve a list of subscriptions with filtering and pagination.
      parameters:
        - name: customerId
          in: query
          description: Filter by customer ID
          schema:
            type: string
        - name: status
          in: query
          description: Filter by subscription status
          schema:
            $ref: '#/components/schemas/SubscriptionStatus'
        - name: page
          in: query
          description: Page number for pagination
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          description: Number of items per page
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Subscriptions retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubscriptionListResponse'
...
  /analytics/transactions:
    post:
      tags:
        - Analytics
      summary: Generate transaction report
      description: |
        Generate detailed transaction reports with advanced filtering options,
        pagination, and export capabilities.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransactionReportRequest'
      responses:
        '200':
          description: Transaction report generated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionReportResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /auth/login endpoint
    
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for service-to-service communication
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      description: Unique key to prevent duplicate processing of the same request
      schema:
        type: string
        maxLength: 255
      example: "unique-key-12345"
    CorrelationId:
      name: X-Correlation-ID
      in: header
      description: Unique identifier for request tracing across systems
      schema:
        type: string
        maxLength: 100
      example: "corr-123e4567-e89b-12d3-a456-426614174000"
  schemas:
...
    LoginRequest:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
          description: Username or email address
          example: john@example.com
        password:
          type: string
          description: User password
          format: password
          example: securePassword123
...
    AuthResponse:
      type: object
      properties:
        accessToken:
          type: string
          description: JWT access token
        refreshToken:
          type: string
          description: Refresh token for token renewal
        tokenType:
          type: string
          example: Bearer
        expiresIn:
          type: integer
          description: Token expiration time in seconds
          example: 3600
        user:
          $ref: '#/components/schemas/UserInfo'
...
    PurchaseRequest:
      type: object
      required:
        - amount
        - currency
        - paymentMethod
        - customer
      properties:
        amount:
          type: number
          format: decimal
          minimum: 0.01
          example: 99.99
        currency:
          type: string
          minLength: 3
          maxLength: 3
          example: USD
        paymentMethod:
          $ref: '#/components/schemas/PaymentMethodRequest'
        customer:
          $ref: '#/components/schemas/CustomerRequest'
        billingAddress:
          $ref: '#/components/schemas/AddressRequest'
        shippingAddress:
          $ref: '#/components/schemas/AddressRequest'
        order:
          $ref: '#/components/schemas/OrderRequest'
        metadata:
          type: object
          additionalProperties: true
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error category
          example: VALIDATION_ERROR
        message:
          type: string
          description: Human-readable error message
          example: Invalid payment method
        details:
          type: string
          description: Detailed error information
        correlationId:
          type: string
          description: Request correlation ID for tracing
          example: corr-123e4567-e89b-12d3-a456-426614174000
        timestamp:
          type: string
          format: date-time
        path:
          type: string
          description: API endpoint path
          example: /api/v1/payments/purchase
...
  responses:
    BadRequest:
      description: Bad request - invalid input parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: VALIDATION_ERROR
            message: Invalid request parameters
            details: Amount must be greater than 0
            correlationId: corr-123e4567-e89b-12d3-a456-426614174000
            timestamp: "2025-09-10T14:30:00Z"
            path: /api/v1/payments/purchase
    Unauthorized:
      description: Unauthorized - invalid or missing authentication
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: AUTHENTICATION_ERROR
            message: Invalid or expired token
            correlationId: corr-123e4567-e89b-12d3-a456-426614174000
            timestamp: "2025-09-10T14:30:00Z"
...
    RateLimited:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: RATE_LIMIT_EXCEEDED
            message: Too many requests
            details: Rate limit of 1000 requests per hour exceeded
            correlationId: corr-123e4567-e89b-12d3-a456-426614174000
            timestamp: "2025-09-10T14:30:00Z"
      headers:
        X-RateLimit-Limit:
          description: Request limit per hour
          schema:
            type: integer
            example: 1000
        X-RateLimit-Remaining:
          description: Remaining requests in current window
          schema:
            type: integer
            example: 0
        X-RateLimit-Reset:
          description: Time when rate limit resets
          schema:
            type: integer
            example: 1631234567
...